Starting code
This code loads the two datasets, removes redundant columns, and merges the two datasets. It also attempts to binarize the columns with multiple attributes (like the genres column) to prepare for scikit learn, but runs into an error because it thinks that there are null values that it can't read.